Removed UploadImageActivity and added the uploading straight to the fragment instead. RecyclerView still not done
